fix: remove unused hook
