+++ src/utils/testing/useConditionalAutoClick.tsx
@@ -1,33 +1,45 @@
 import { useEffect } from "react";
 
-interface UseAutoClickElementOptions {
+interface ConditionConfig {
   /**
-   * The CSS selector for the element to click
+   * Query parameter to check for in the URL
+   * @default "autoClick"
+   */
+  queryParam?: string;
+  /**
+   * When to trigger the auto-click
+   * @default "present"
+   */
+  when?: "present" | "absent";
+}
+
+interface UseConditionalAutoClickProps {
+  /**
+   * CSS selector for the element to click
    */
   selector: string;
-
+  /**
+   * Condition configuration for when to auto-click
+   */
+  condition?: ConditionConfig;
   /**
    * Delay in milliseconds before attempting to click the element
    * @default 100
    */
   delay?: number;
-
   /**
    * Whether to log the action to the console
    * @default true
    */
   logging?: boolean;
-
   /**
    * Callback to execute when the element is found and clicked
    */
   onSuccess?: () => void;
-
   /**
    * Callback to execute when the element is not found
    */
   onNotFound?: () => void;
-
   /**
    * Callback to execute when an error occurs
    */
@@ -35,30 +47,39 @@ interface UseAutoClickElementOptions {
 }
 
 /**
- * A hook that automatically clicks on an element once it's available in the DOM
+ * Hook that conditionally auto-clicks an element based on URL query parameters.
+ * Useful for preventing double-clicks during automated testing.
  *
  * @example
- * // Click on the app switcher once the component mounts
- * useAutoClickElement({ selector: 'button[aria-label="App Switcher"]' });
- *
- * @example
- * // Click with a longer delay and custom callbacks
- * useAutoClickElement({
- *   selector: '.custom-dropdown-trigger',
- *   delay: 500,
- *   onSuccess: () => console.log('Dropdown opened!'),
- *   onNotFound: () => console.warn('Could not find dropdown trigger')
+ * // Only auto-click when NOT in Cypress testing
+ * useConditionalAutoClick({
+ *   selector: 'button[aria-label="Menu"]',
+ *   condition: {
+ *     queryParam: "cypressTest",
+ *     when: "absent"
+ *   }
  * });
  */
-export function useAutoClickElement({
+export default function useConditionalAutoClick({
   selector,
+  condition = {},
   delay = 100,
   logging = true,
   onSuccess,
   onNotFound,
   onError,
-}: UseAutoClickElementOptions): void {
+}: UseConditionalAutoClickProps): void {
+  const { queryParam = "autoClick", when = "present" } = condition;
+
+  // Check if the condition is met
+  const urlParams = new URLSearchParams(window.location.search);
+  const paramExists = urlParams.has(queryParam);
+  // Auto-click when param exists and when="present", or when param doesn't exist and when="absent"
+  const shouldAutoClick = (paramExists && when === "present") || (!paramExists && when === "absent");
+
   useEffect(() => {
+    if (!shouldAutoClick) return;
+
     const timeoutId = setTimeout(() => {
       try {
         const element = document.querySelector(selector) as HTMLElement;
@@ -83,7 +104,5 @@ export function useAutoClickElement({
     }, delay);
 
     return () => clearTimeout(timeoutId);
-  }, [selector, delay, logging, onSuccess, onNotFound, onError]);
+  }, [selector, delay, logging, onSuccess, onNotFound, onError, shouldAutoClick]);
 }
-
-export default useAutoClickElement;
